requirements.txt
kaggle>=1.5.13
pandas>=1.5.0
google-genai